+++ packages/sdk/src/StreamrClient.ts
@@ -32,7 +32,6 @@ import { RpcProviderSource } from './RpcProviderSource'
 import { Stream } from './Stream'
 import { StreamIDBuilder } from './StreamIDBuilder'
 import { StreamMetadata, getPartitionCount } from './StreamMetadata'
-import { StreamrClientError } from './StreamrClientError'
 import { ContractFactory } from './contracts/ContractFactory'
 import { Operator } from './contracts/Operator'
 import { OperatorRegistry } from './contracts/OperatorRegistry'
@@ -175,9 +174,6 @@ export class StreamrClient {
     async updateEncryptionKey(opts: UpdateEncryptionKeyOptions): Promise<void> {
         if (opts.streamId === undefined) {
             throw new Error('streamId required')
-        }
-        if (opts.key !== undefined && this.config.encryption.litProtocolEnabled) {
-            throw new StreamrClientError('cannot pass "key" when Lit Protocol is enabled', 'UNSUPPORTED_OPERATION')
         }
         const streamId = await this.streamIdBuilder.toStreamID(opts.streamId)
         const queue = await this.publisher.getGroupKeyQueue(streamId)