+++ packages/sdk/src/encryption/GroupKeyManager.ts
@@ -1,4 +1,4 @@
-import { StreamID, StreamPartID, StreamPartIDUtils, UserID, waitForEvent } from '@streamr/utils'
+import { StreamID, StreamPartID, UserID, waitForEvent } from '@streamr/utils'
 import crypto from 'crypto'
 import { Lifecycle, inject, scoped } from 'tsyringe'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
@@ -7,7 +7,6 @@ import { DestroySignal } from '../DestroySignal'
 import { StreamrClientEventEmitter } from '../events'
 import { uuid } from '../utils/uuid'
 import { GroupKey } from './GroupKey'
-import { LitProtocolFacade } from './LitProtocolFacade'
 import { LocalGroupKeyStore } from './LocalGroupKeyStore'
 import { SubscriberKeyExchange } from './SubscriberKeyExchange'
 
@@ -15,7 +14,6 @@ import { SubscriberKeyExchange } from './SubscriberKeyExchange'
 export class GroupKeyManager {
 
     private readonly subscriberKeyExchange: SubscriberKeyExchange
-    private readonly litProtocolFacade: LitProtocolFacade
     private readonly localGroupKeyStore: LocalGroupKeyStore
     private readonly config: Pick<StrictStreamrClientConfig, 'encryption'>
     private readonly authentication: Authentication
@@ -24,7 +22,6 @@ export class GroupKeyManager {
 
     constructor(
         subscriberKeyExchange: SubscriberKeyExchange,
-        litProtocolFacade: LitProtocolFacade,
         localGroupKeyStore: LocalGroupKeyStore,
         @inject(ConfigInjectionToken) config: Pick<StrictStreamrClientConfig, 'encryption'>,
         @inject(AuthenticationInjectionToken) authentication: Authentication,
@@ -32,7 +29,6 @@ export class GroupKeyManager {
         destroySignal: DestroySignal
     ) {
         this.subscriberKeyExchange = subscriberKeyExchange
-        this.litProtocolFacade = litProtocolFacade
         this.localGroupKeyStore = localGroupKeyStore
         this.config = config
         this.authentication = authentication
@@ -41,24 +37,13 @@ export class GroupKeyManager {
     }
 
     async fetchKey(streamPartId: StreamPartID, groupKeyId: string, publisherId: UserID): Promise<GroupKey> {
-        const streamId = StreamPartIDUtils.getStreamID(streamPartId)
-
         // 1st try: local storage
         let groupKey = await this.localGroupKeyStore.get(groupKeyId, publisherId)
         if (groupKey !== undefined) {
             return groupKey
         }
 
-        // 2nd try: lit-protocol
-        if (this.config.encryption.litProtocolEnabled) {
-            groupKey = await this.litProtocolFacade.get(streamId, groupKeyId)
-            if (groupKey !== undefined) {
-                await this.localGroupKeyStore.set(groupKey.id, publisherId, groupKey.data)
-                return groupKey
-            }
-        }
-
-        // 3rd try: Streamr key-exchange
+        // 2nd try: Streamr key-exchange
         await this.subscriberKeyExchange.requestGroupKey(groupKeyId, publisherId, streamPartId)
         const groupKeyIds = await waitForEvent(
             // TODO remove "as any" type casing in NET-889
@@ -86,13 +71,7 @@ export class GroupKeyManager {
         }
         if (groupKey === undefined) {
             const keyData = crypto.randomBytes(32)
-            // 1st try lit-protocol, if a key cannot be generated and stored, then generate group key locally
-            if (this.config.encryption.litProtocolEnabled) {
-                groupKey = await this.litProtocolFacade.store(streamId, keyData)
-            }
-            if (groupKey === undefined) {
-                groupKey = new GroupKey(uuid('GroupKey'), keyData)
-            }
+            groupKey = new GroupKey(uuid('GroupKey'), keyData)
         }
         await this.localGroupKeyStore.set(groupKey.id, publisherId, groupKey.data)
         await this.localGroupKeyStore.setLatestEncryptionKeyId(groupKey.id, publisherId, streamId)